fix(compiler): cast reduce accumulator initial value to Int64(0) in ConversionDiagnosticLogger
- ConversionDiagnosticLogger.swift: cast reduce accumulator initial value to Int64(0) to fix Int64 type mismatch.
- CBZToEPUBConverter, EPUBMerger, PDFGenerator: change var diagMetrics to let diagMetrics.

+++ ComicToPDF/ComicToPDF/Services/Conversion/ConversionDiagnosticLogger.swift
@@ -29,7 +29,7 @@ struct ConversionDiagnosticLogger {
     /// Call at the start of any conversion process (single file, convert & merge, omnibus, etc.)
     static func logStart(jobTitle: String, settings: ConversionSettings, sourceFiles: [URL]) -> ConversionMetrics {
         let preset = settings.compressionQuality
-        let totalInputBytes = sourceFiles.reduce(0) { sum, url in
+        let totalInputBytes = sourceFiles.reduce(Int64(0)) { sum, url in
             let size = (try? FileManager.default.attributesOfItem(atPath: url.path)[.size] as? UInt64) ?? 0
             return sum + Int64(size)
         }
